pyproject.toml
[tool.poetry]
name = "mfa-stablediffusion"
version = "0.1.0"
description = "stable diffusion wrapper for MFA"
authors = ["Alex Scotland"]
packages = [{ include = "mfa-stablediffusion" }]

[tool.poetry.dependencies]
python = "^3.9" # Keep
modular-fastapi = {path = "/app/api/"}  # Uses main API’s FastAPI version
requests = "^2.31.0"  # Plugin-specific dependency
diffusers = "0.31.0"
torch = "2.3.0"
transformers = "4.41.0"
tokenizers ="0.19.1"
accelerate = "0.30.1"
safetensors = "0.4.3"
peft = "0.11.1"
opencv-python = "4.10.0.82"
python-multipart = "0.0.9"
pyyaml = "6.0.2"
setuptools = "77.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
